refactor: port DSO instrument layer to Pico ADC
