+++ src/application/dso_commands.c
@@ -1,6 +1,6 @@
 #include "application/dso_commands.h"
 
-#include "system/dso.h"
+#include "system/instrument/dso.h"
 
 void dso_commands_reset(void) { dso_reset_state(); }
